Use YAML for database settings
Replaces shell environment variables and Embedded Ruby (ERB). This
will allow easier merges of upstream `project.yaml` files, as well
as allowing for customization of database settings.

It is common for database settings for production and development
to be different, because of possibly different hosts for OSM data
and CanVec data. This is different from openstreetmap-carto which
only uses a single database.