Move database projection to EPSG:3573
This fixes *a lot* of problems with text rendering. Clipping on
maritime boundaries is fixed. Text labels for regions (e.g.
Baffin Bay) now are correctly shown. ("Canada" is still missing
but that is because of the source extract.)

Rendering should also be faster with the OSM data not needing to
be re-projected by Mapnik from EPSG:4326 to EPSG:3573.

This does mean that the database cannot be re-used for other
projections (unless they are very similar to EPSG:3573).