Add bias to northern place name rendering
Massively increases rendering placement score for communities
north of 60˚N. For places in the bounding box, they are
classified as cities, villages are included, and their population
score is multiplied by 100.

I did not use any special method to determine these numbers, only
what looked decent on the preview in Kosmtik.

Currently uses bounding box for [-180, 60, 0, 90] but that could
be changed to a simple polygon.